+++ lib/Controller/ExportController.php
@@ -21,7 +21,7 @@ use \OCP\AppFramework\Http\JSONResponse;
 use OCP\IUserSession;
 use OCP\AppFramework\Http\Attribute\NoAdminRequired;
 use OCP\AppFramework\Http\Attribute\NoCSRFRequired;
-use OCP\AppFramework\Http\Attribute\CORS;
+use OCP\AppFramework\Http\Attribute\FrontpageRoute;
 
 /**
  * Class ExportController
@@ -43,6 +43,7 @@ class ExportController extends Controller
 
     #[NoCSRFRequired]
     #[NoAdminRequired]
+    #[FrontpageRoute(verb: 'GET', url: '/export/opml')]
     public function opml(): DataDownloadResponse
     {
         $date = date('Y-m-d');
@@ -55,6 +56,7 @@ class ExportController extends Controller
     }
 
     #[NoAdminRequired]
+    #[FrontpageRoute(verb: 'GET', url: '/export/articles')]
     public function articles(): JSONResponse
     {
         $articles = $this->exportService->articles($this->getUserId());