+++ lib/Controller/FaviconApiController.php
@@ -13,11 +13,14 @@ use OCP\IUserSession;
 use OCP\AppFramework\Http\Attribute\CORS;
 use OCP\AppFramework\Http\Attribute\NoAdminRequired;
 use OCP\AppFramework\Http\Attribute\NoCSRFRequired;
+use OCP\AppFramework\Http\Attribute\ApiRoute;
+use OCP\AppFramework\Http\Attribute\OpenAPI;
 use OCP\AppFramework\Http\DataDownloadResponse;
 
 use OCA\News\Constants;
 use OCA\News\Utility\AppData;
 
+#[OpenAPI(scope: OpenAPI::SCOPE_DEFAULT)]
 class FaviconApiController extends ApiController
 {
     public function __construct(
@@ -36,6 +39,7 @@ class FaviconApiController extends ApiController
     #[CORS]
     #[NoCSRFRequired]
     #[NoAdminRequired]
+    #[ApiRoute(verb: 'GET', url: '/api/{apiVersion}/favicon/{feedUrlHash}', requirements: ['apiVersion' => 'v1-3'])]
     public function get(string $feedUrlHash)
     {
         $feed_logo = null;