+++ lib/Controller/ItemController.php
@@ -20,6 +20,7 @@ use OCP\AppFramework\Http\JSONResponse;
 use \OCP\IRequest;
 use \OCP\AppFramework\Http;
 use OCP\AppFramework\Http\Attribute\NoAdminRequired;
+use OCP\AppFramework\Http\Attribute\FrontpageRoute;
 use OCP\Config\IUserConfig;
 
 use \OCA\News\Service\Exceptions\ServiceException;
@@ -64,6 +65,7 @@ class ItemController extends Controller
      * @return array
      */
     #[NoAdminRequired]
+    #[FrontpageRoute(verb: 'GET', url: '/items')]
     public function index(
         int $type = 3,
         int $id = 0,
@@ -185,6 +187,7 @@ class ItemController extends Controller
      * @return array
      */
     #[NoAdminRequired]
+    #[FrontpageRoute(verb: 'GET', url: '/items/new')]
     public function newItems(int $type, int $id, $lastModified = 0): array
     {
         $showAll = $this->userConfig->getValueString(
@@ -246,6 +249,7 @@ class ItemController extends Controller
      * @return array|JSONResponse
      */
     #[NoAdminRequired]
+    #[FrontpageRoute(verb: 'POST', url: '/items/{feedId}/{guidHash}/star')]
     public function star(int $feedId, string $guidHash, bool $isStarred)
     {
         try {
@@ -270,6 +274,7 @@ class ItemController extends Controller
      * @return array|JSONResponse
      */
     #[NoAdminRequired]
+    #[FrontpageRoute(verb: 'POST', url: '/items/{itemId}/read')]
     public function read(int $itemId, $isRead = true)
     {
         try {
@@ -288,6 +293,7 @@ class ItemController extends Controller
      * @return array
      */
     #[NoAdminRequired]
+    #[FrontpageRoute(verb: 'POST', url: '/items/read')]
     public function readAll(int $highestItemId): array
     {
         $this->itemService->readAll($this->getUserId(), $highestItemId);
@@ -301,6 +307,7 @@ class ItemController extends Controller
      * @return void
      */
     #[NoAdminRequired]
+    #[FrontpageRoute(verb: 'POST', url: '/items/read/multiple')]
     public function readMultiple(array $itemIds): void
     {
         foreach ($itemIds as $id) {
@@ -318,6 +325,7 @@ class ItemController extends Controller
      * @param string $shareRecipientId User to share the item with
      */
     #[NoAdminRequired]
+    #[FrontpageRoute(verb: 'POST', url: '/items/{itemId}/share/{shareRecipientId}')]
     public function share(int $itemId, string $shareRecipientId)
     {
         try {